Remove wsrep-new-cluster from cluster.cnf
During upgrade, the wsrep-new-cluster is added to the
cluster.cnf file as a fall back in case the mysql bootstrap
command does not work.  The problem with this is that node1
always attempts to start a new cluster which causes issues during
a Galera upgrade or a rolling restart as node1 refuses to
come on-line and join the other nodes in the cluster.

This commit removes the value from the configuration file after
bootstrap so that it's only used during the playbook if needed
and isn't around for production operations.

Change-Id: Ib3ad0dc58352cabb6157a98d1d12fa788b4a0a68